Moves shared geocode functionality to a module.
